requirements.txt
# python3.8
docopt
pyyaml